+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 # Changed
 - Updated description for `Gifted` feature.
 - Renamed `Auto Metamagic` to `Overpowered Ability — Auto Metamagic`.
+- Renamed additional `Signature Move` and `Beach Episode` selections for `Isekai Protagonist`.
 
 # [5.0.4] - 2023-04-30
 # Fixed